Restructure mimuMajorityVoteAlg Types.h to pure C
